+++ test/org/cyverse/oai_ore_test.clj
@@ -11,23 +11,29 @@
 ;; URIs to use when building test OREs.
 (def ^:private agg-uri "http://foo.org")
 (def ^:private arch (file-info-from-id "6cf74bf2-f172-47b4-94cc-acfc82559fc6"))
+(def ^:private metadata (file-info-from-id "d0b7aa9a-1c22-4a59-9b04-3f35f3a9e15d"))
 (def ^:private archived-files
   (mapv file-info-from-id ["73cd1d79-28c9-4a52-9060-cf9c65905922" "6772f358-9c10-4fa0-9fec-1376816b37e4"]))
 
 ;; Comparing generated XML to parsed XML requires us to serialize and parse the generated XML.
-(defn- test-ore [filename ore]
+(defn- test-ore [filename ore & [debug]]
+  (when debug
+    (println "Actual:" (xml/indent-str (to-rdf ore)))
+    (println "Expected:" (xml/indent-str (xml/parse (io/reader (io/resource filename))))))
   (is (= (xml/parse (io/reader (io/resource filename)))
          (xml/parse-str (xml/emit-str (to-rdf ore))))))
 
 (deftest test-empty-ore
   (testing "Empty ORE."
-    (test-ore "empty-ore.rdf" (build-ore agg-uri arch []))
-    (test-ore "empty-ore.rdf" (build-ore agg-uri arch [] [{:attr "datacite.title" :value ""}]))))
+   (test-ore "empty-ore.rdf" (build-ore agg-uri arch []))
+   (test-ore "empty-ore.rdf" (build-ore agg-uri arch [] [{:attr "datacite.title" :value ""}]))
+   (test-ore "empty-ore-meta.rdf" (build-ore agg-uri arch [] [] metadata))))
 
 (deftest test-ore-with-files
   (testing "ORE with files."
     (test-ore "ore-with-files.rdf" (build-ore agg-uri arch archived-files))
-    (test-ore "ore-with-files.rdf" (build-ore agg-uri arch archived-files [{:attr "datacite.title" :value ""}]))))
+    (test-ore "ore-with-files.rdf" (build-ore agg-uri arch archived-files [{:attr "datacite.title" :value ""}]))
+    (test-ore "ore-with-files-meta.rdf" (build-ore agg-uri arch archived-files [] metadata))))
 
 (deftest test-ore-with-title
   (testing "ORE with title."